// server/types.ts
export interface User {
  id: number
  uuid: string
  email: string
  password: string
  avatar?: string
  first_name: string
  last_name: string
  email_verified?: boolean
  is_active?: boolean
  last_login?: Date
  phone?: string
  created_at: Date
  deleted_at?: Date
  updated_at?: Date
  role_id: number
  metadata?: JSON
  notifications?: JSON
  status?: UserStatus
  role: Role
  twofactor?: TwoFactor
  author?: Author
  chats: Chat[]
  notification: Notification[]
  posts: Post[]
  comments: Comment[]
}
// ...
export interface Wallet {
  id: number
  uuid: string
  type: WalletType
  user_id: number
  balance: number
  inOrder: number
  currency: string
  addresses: any
  created_at: string
  updated_at: string
  transactions: Transaction[]
  invoices: Invoice[]
}

export enum WalletType {
  FIAT = 'FIAT',
  SPOT = 'SPOT',
  ECO = 'ECO',
}
// ...
export enum TransactionType {
  FAILED = 'FAILED',
  DEPOSIT = 'DEPOSIT',
  WITHDRAW = 'WITHDRAW',
  OUTGOING_TRANSFER = 'OUTGOING_TRANSFER',
  INCOMING_TRANSFER = 'INCOMING_TRANSFER',
  PAYMENT = 'PAYMENT',
  REFUND = 'REFUND',
  BINARY_ORDER = 'BINARY_ORDER',
  EXCHANGE_ORDER = 'EXCHANGE_ORDER',
  INVESTMENT = 'INVESTMENT',
  INVESTMENT_ROI = 'INVESTMENT_ROI',
  AI_INVESTMENT = 'AI_INVESTMENT',
  AI_INVESTMENT_ROI = 'AI_INVESTMENT_ROI',
  INVOICE = 'INVOICE',
  FOREX_DEPOSIT = 'FOREX_DEPOSIT',
  FOREX_WITHDRAW = 'FOREX_WITHDRAW',
  FOREX_INVESTMENT = 'FOREX_INVESTMENT',
  FOREX_INVESTMENT_ROI = 'FOREX_INVESTMENT_ROI',
  ICO_CONTRIBUTION = 'ICO_CONTRIBUTION',
  REFERRAL_REWARD = 'REFERRAL_REWARD',
  STAKING = 'STAKING',
  STAKING_REWARD = 'STAKING_REWARD',
  P2P_OFFER_TRANSFER = 'P2P_OFFER_TRANSFER',
  P2P_TRADE = 'P2P_TRADE',
}

export enum TransactionStatus {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED',
  REJECTED = 'REJECTED',
  REFUNDED = 'REFUNDED',
  TIMEOUT = 'TIMEOUT',
}

// The Transaction type maps to your Prisma model for transactions
export interface Transaction {
  id: number
  uuid: string
  user_id: number
  wallet_id: number
  type: TransactionType
  status: TransactionStatus
  amount: number
  fee?: number
  description?: string
  metadata?: any // Replace with a more specific type if you know the shape of the metadata
  reference_id?: string
  created_at: Date
  updated_at?: Date
  wallet: Wallet
  user: User
  invoice: Invoice
}

export interface Invoice {
  id: number
  wallet_id: number
  user_id: number
  amount: number
  status: string
  created_at: string
  updated_at: string
}
// ...
export enum StakingStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  COMPLETED = 'COMPLETED',
}

export enum StakeStatus {
  ACTIVE = 'ACTIVE',
  RELEASED = 'RELEASED',
  WITHDRAWN = 'WITHDRAWN',
}

export enum RewardStatus {
  PENDING = 'PENDING',
  DISTRIBUTED = 'DISTRIBUTED',
}

// Model Types
export type StakingPool = {
  id: number
  uuid: string
  name: string
  description: string
  currency: string
  chain: string
  type: WalletType
  min_stake: number
  max_stake: number
  status: StakingStatus
  durations: StakingDuration[]
  stakes: StakingLog[]
  created_at: Date
}

export type StakingDuration = {
  id: number
  pool_id: number
  duration: number // Duration in days
  interest_rate: number // Annual interest rate as a percentage
  pool: StakingPool
  created_at: Date
}

export type StakingLog = {
  id: number
  uuid: string
  user_id: number
  pool_id: number
  amount: number // Amount of tokens staked
  stake_date: Date // Date when the stake was made
  release_date: Date // Date when the stake can be withdrawn
  status: StakeStatus
  rewards: StakingReward[]
  pool: StakingPool
  transaction: Transaction
  created_at: Date
}

export type StakingReward = {
  id: number
  stake_id: number
  amount: number // Reward amount
  reward_date: Date // Date when the reward was distributed
  status: RewardStatus
  stake: StakingLog
  created_at: Date
}